requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
PyMySQL==1.1.2
python-decouple==3.8
sqlparse==0.5.4
tzdata==2025.2
